requirements_local_model.txt
# 本地模型部署所需依赖
# 替代 moonshot-v1-8k-vision-preview 的本地化方案

# ============= 核心依赖 =============
# YOLO 模型框架（推荐）
ultralytics>=8.0.0

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# ============= 可选：GPU 加速 =============
# 如果有 NVIDIA GPU，建议安装 CUDA 版本的 PyTorch
# 访问 https://pytorch.org/ 选择对应的 CUDA 版本
# 例如 CUDA 11.8:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118

# ============= 图像处理 =============
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# ============= 可选：其他视觉模型 =============
# 如果想使用其他模型，可以安装以下库：

# CLIP (OpenAI 的图像-文本模型)
# git+https://github.com/openai/CLIP.git

# Detectron2 (Facebook 的目标检测框架)
# 需要从源码安装，参考: https://github.com/facebookresearch/detectron2

# Hugging Face Transformers (支持多种视觉模型)
# transformers>=4.30.0
# timm>=0.9.0

# ============= 工具库 =============
# 进度条
tqdm>=4.65.0

# 日志
loguru>=0.7.0
